backend/requirements.txt
pip~=23.3.1
pytz~=2023.3.post1
asgiref~=3.7.2
sqlparse~=0.4.4
setuptools~=68.2.2
Django~=4.2.6
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
django-mysql~=4.12.0
mysqlclient~=2.2.0
